reorganize utils
