+++ src/components/AppSelector.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import { Check, X } from 'lucide-react'
 import { Command, CommandEmpty, CommandGroup, CommandItem, CommandList } from '@/components/ui/command'
 import { Badge } from '@/components/ui/badge'
-import { cn } from '@/lib/tailwind.util'
+import { cn } from '@/lib/utils/tailwind.util'
 import { AppCategory, AppDefinition, categoryEmojis } from '@/lib/app-directory/apps-types'
 import { apps } from '@/lib/app-directory/apps-list'
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip'
