+++ src/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { cn } from '@/lib/tailwind.util'
+import { cn } from '@/lib/utils/tailwind.util'
 
 const Card = React.forwardRef<
   HTMLDivElement,
